feat: 일반적 로직을 General 로 분리
