+++ src/script/parser/generalParser.ts
@@ -1,30 +1,7 @@
-import fs from 'fs';
 import { convertKeysToUpperCase, isEmptyObject } from '../../utils/my_lodash';
+import { General } from '../general';
 
-export class GeneralParser {
-  static readFileSync(path: string) {
-    return fs.readFileSync(path, 'utf8');
-  }
-
-  static async writeFile(path: string, data: string) {
-    const directory = path.split('/').slice(0, -1).join('/');
-    if (!fs.existsSync(path)) {
-      await fs.promises.mkdir(directory, { recursive: true });
-    }
-    await fs.promises.writeFile(path, data);
-  }
-
-  /**
-   * JSON string 데이터를 구조화된 자바스크립트 value, object 로 변환
-   */
-  static parse(data: string) {
-    return JSON.parse(data);
-  }
-
-  static applyRegex(target: string, ...args: ((text: string) => string)[]) {
-    return args.reduce((acc, fn) => fn(acc), target);
-  }
-
+export class GeneralParser extends General {
   static replaceVariables(text: string, variables: any) {
     if (isEmptyObject(variables)) return text;
     convertKeysToUpperCase(variables);
